+++ CHANGES.md
@@ -4,6 +4,7 @@ master (in development)
 - Add `flycheck-pos-tip-timeout` [GH-9]
 - Fall back to echo area display of errors on TTY frames [GH-11] [GH-14]
 - Remove `flycheck-pos-tip-show-function`
+- Add `flycheck-pos-tip-display-errors-tty-function`
 
 0.1 (Nov 16, 2015)
 ==================